requirements.txt
streamlit>=1.20
langchain>=0.0.300
openai>=1.0.0
chromadb>=0.3.26
pypdf>=3.14.0
python-docx>=0.8.11
docx2txt>=0.8
pandas>=2.0
tqdm>=4.65
python-dotenv>=1.0
requests>=2.31
sentence-transformers>=2.2.2
